+++ st2tests/integration/orquesta/test_wiring_with_items.py
@@ -287,3 +287,10 @@ class WithItemsWiringTest(base.TestWorkflowExecution):
 
         # Wait for completion.
         ex = self._wait_for_state(ex, ac_const.LIVEACTION_STATUS_SUCCEEDED)
+
+    def test_subworkflow_empty_with_items(self):
+        wf_name = 'examples.orquesta-test-subworkflow-empty-with-items'
+        ex = self._execute_workflow(wf_name)
+        ex = self._wait_for_completion(ex)
+
+        self.assertEqual(ex.status, ac_const.LIVEACTION_STATUS_SUCCEEDED)